Fix type of variable

+++ libft/ft_strlen.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mrabelo- <mrabelo-@student.42berlin.d      +#+  +:+       +#+        */
+/*   By: mrabelo- <mrabelo-@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/13 15:27:26 by mrabelo-          #+#    #+#             */
-/*   Updated: 2023/11/13 17:20:39 by mrabelo-         ###   ########.fr       */
+/*   Updated: 2023/11/23 15:11:41 by mrabelo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char*s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
@@ -27,6 +27,6 @@ size_t	ft_strlen(const char*s)
 int	main()
 {
 	int	test1 = ft_strlen("helloworld");
-	
+
 	printf("%i\n", test1);
 }*/
